move all components in Home(should be Search) page to seperate files
