+++ ubg/src/common/Navbar.js
@@ -6,7 +6,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import IconButton from '@material-ui/core/IconButton';
 import Typography from '@material-ui/core/Typography';
 import AuthButtons from './AuthButtons';
-import Search from './Search';
+import SearchField from './SearchField';
 
 const useStyles = makeStyles((theme) => ({
     menuButton: {
@@ -28,7 +28,7 @@ export default function Navbar() {
           <Typography variant="h6" className={classes.title}>
             UltimateBoardGame
           </Typography>
-          <Search />
+          <SearchField />
           <AuthButtons />
         </Toolbar>
       </AppBar>
